Deploy configuration added, Stat and Links improvement

+++ src/links/views.py
@@ -1,30 +1,41 @@
+from datetime import timedelta
+
+from django.contrib.auth.mixins import LoginRequiredMixin
 from django.db.models import ProtectedError
-from django.http import JsonResponse, HttpResponse
-from django.shortcuts import render, get_object_or_404
-
-from hashlib import sha256
-
-from django.views.generic import ListView, RedirectView, CreateView, DeleteView
+from django.http import Http404, HttpResponse, JsonResponse
+from django.shortcuts import get_object_or_404, render
+from django.urls import reverse
+from django.utils import timezone
+from django.views import View
+from django.views.generic import CreateView, DetailView, ListView, RedirectView
 
 from links.forms import LinkForm
-from links.models import Link, generate_hash_salt
-from base64 import b64encode
+from links.models import Link, generate_short_link
 
 
-class LinkListView(ListView):
+class LinkListView(LoginRequiredMixin, ListView):
     """
-    GET request url: <server>/links/list/
+    GET request
+    url: <server>/links/list/
     Returns:
         Rendered template links_list.html with links.
-    # TODO update template page layout.
     """
 
+    login_url = "/auth/sign_in/"
+    redirect_field_name = "redirect_to"
     model = Link
-    template_name = "link_list.html"
+    template_name = "links/list.html"
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         return context
+
+    def get_queryset(self):
+        user = self.request.user
+        links = Link.objects.filter(
+            creator=user,
+        )
+        return links
 
 
 class LinkRedirectView(RedirectView):
@@ -34,7 +45,8 @@ class LinkRedirectView(RedirectView):
 
     def get_redirect_url(self, *args, **kwargs):
         """
-        GET request url: <server>/<slug:short_link>/
+        GET request
+        url: <server>/<slug:short_link>/
         Args:
             request: Django Request object.
             *args: Positional args.
@@ -43,18 +55,23 @@ class LinkRedirectView(RedirectView):
         Redirects to origin url and , or send 404 http error if short_link doesn't exist.
         """
         link = get_object_or_404(Link, short_link=kwargs["short_link"])
-        link.views += 1
-        link.save()
-        return link.origin_url
+        if link.permanent or link.expire_time >= timezone.now():
+            link.views.create(date=timezone.now().date())
+            link.save()
+            return link.origin_url
+        raise Http404
 
 
-# TODO LinkDeleteView.
-class LinkDeleteView(DeleteView):
-    model = Link
+class LinkDeleteView(LoginRequiredMixin, View):
 
-    def delete(self, request, *args, **kwargs):
+    login_url = "/auth/sign_in/"
+    redirect_field_name = "redirect_to"
+    http_method_names = ["post"]
+
+    def post(self, request, *args, **kwargs):
         """
-        DELETE request url: <server>/links/<slug:short_url/
+        POST request
+        url: <server>/links/<slug:short_url/
         Args:
             request: Django Request object.
             *args: Positional args.
@@ -63,20 +80,24 @@ class LinkDeleteView(DeleteView):
         Returns:
             HttpResponse with status code 200 if deleted successfully or 404
         """
-        link = get_object_or_404(self.model, short_link=kwargs["short_url"])
-        try:
-            link.delete()
-        except ProtectedError:
-            return HttpResponse("Link can't be deleted.", status=400)
-        return HttpResponse(status=200)
+        link = get_object_or_404(Link, id=kwargs["pk"])
+        if link.creator.email == request.user.email:
+            try:
+                link.delete()
+                return HttpResponse(reverse("links:list"), status=200)
+            except ProtectedError:
+                return HttpResponse("Link can't be deleted.", status=400)
+        else:
+            return HttpResponse(status=403)
 
 
 class LinkCreateView(CreateView):
-    create_template = "index.html"
+    create_template = "links/create.html"
 
     def get(self, request, *args, **kwargs):
         """
-        GET request url: <server>/links/
+        GET request
+        url: <server>/links/
         Args:
             request: Django Request object.
             *args: Positional args.
@@ -89,27 +110,46 @@ class LinkCreateView(CreateView):
 
     def post(self, request, *args, **kwargs):
         """
-        POST request to create short link. url: <server>/links/
+        POST request to create short link.
+        url: <server>/links/
         Args:
             request:
             *args:
             **kwargs:
-
         Returns:
-
+            JsonResponse with link if no errors occurred else message with error
         """
         form = LinkForm(request.POST)
         if form.is_valid():
             _url = form.cleaned_data["url"]  # get url from post data
-            hash_salt = generate_hash_salt()  # generate hash salt
-            hash_func = sha256()  # hashing algorithm
-            hash_func.update((_url + hash_salt).encode("utf-8"))  # hash url with salt
-            short_link = hash_func.hexdigest()  # hash url with salt
-            short_link = b64encode(bytes.fromhex(short_link)).decode()[:10]  # hex to base64 and slice first 10 chars
-            short_link.replace("+", "-")
-            short_link.replace("/", "_")
-            Link.objects.create(origin_url=_url, short_link=short_link)
-            data = {"link": short_link}
+            _short_link = generate_short_link(_url)
+            _expire_time = int(form.cleaned_data["expire_time"])  # get expire_time from post data
+            link = Link.objects.create(origin_url=_url, short_link=_short_link)
+            if _expire_time == 0:
+                link.permanent = True
+            elif _expire_time == 1 or _expire_time == 7:
+                link.expire_time = timezone.now() + timedelta(days=_expire_time)
+            if request.user.is_authenticated:
+                link.creator = request.user  # if user is authenticated add creator to link
+            link.save()
+            data = {"link": _short_link}
             return JsonResponse(data)
-        data = {"error": "error"}
+        data = {"error": form.errors}
         return JsonResponse(data)
+
+
+class LinkDetailView(DetailView):
+    model = Link
+    context_object_name = "link"
+    template_name = "links/detail.html"
+
+    def get_object(self, queryset=None):
+        link = get_object_or_404(Link, creator=self.request.user, id=self.kwargs["pk"])
+        views = link.views.all()
+        views_count_by_date = {}
+        for view in views:
+            if str(view.date) not in views_count_by_date.keys():
+                views_count_by_date[str(view.date)] = 0
+            views_count_by_date[str(view.date)] += 1
+        self.extra_context = {"views": views_count_by_date}
+        return link
